🐛 fix(form_component.dart): refactor inputForm function into a reusable widget MyInputForm ✨ feat(form_component.dart): add support for password visibility toggle in MyInputForm The inputForm function in form_component.dart has been refactored into a reusable widget called MyInputForm. This allows for better code organization and reusability. Additionally, support for a password visibility toggle has been added to MyInputForm, making it easier to handle password input fields.
🐛 fix(sql_helper.dart): add isUsernameExists function to check if a username already exists in the database
The SQLHelper class in sql_helper.dart now includes a static function called isUsernameExists. This function checks if a given username already exists in the user table of the database. It returns true if the username exists and false otherwise. This function is useful for validating unique usernames during user registration.

🐛 fix(login.dart): replace inputForm function with MyInputForm widget
The

🐛 fix(home.dart): import HasilCariNamaKamar to resolve missing import error
✨ feat(home.dart): add search functionality to search for rooms by name
🐛 fix(profile_page.dart): fix Update button text color to white
🐛 fix(register.dart): fix validation function names and add validation for username and email uniqueness
The import statement for `HasilCariNamaKamar` was added to resolve the missing import error. The search functionality was added to allow users to search for rooms by name. The text color of the Update button in `profile_page.dart` was fixed to be white. The validation function names in `register.dart` were fixed and validation for username and email uniqueness was added.